internode: widen NVL gen-stamp slots to 64-bit (seq32, value32)
Replaces C4's 12-bit-seq / 20-bit-value packed `int` slots with full
64-bit `(seq, value)` pairs on `nvl_channel_prefix_start/end`,
`nvl_channel_head`, and `nvl_channel_tail` across dispatch_main,
dispatch_grads_main, and the combine sender + forwarder/coordinator.

Why. The 12-bit seq window (4096 iters) was the substrate for two
failure modes patched by M1's iter-entry force-writes (60ec765,
850fc37§2): (a) sparse-routing dormancy — a substream silent for
≥ 4096 iters whose stale residue then aliases the current seq; (b)
cross-call same-seq aliasing — fwd dispatch and bwd dispatch_grads
share `dispatch_seq` within a layer, so bwd's reader matches fwd's
leftover trivially. A wider window collapses (a) to a non-issue by
construction (4 B iters / ~32-bit seq window vs ~16M iters needed at
production rates × 64 layers × 1M training steps). Case (b) is fixed
in the next commit by a phase bit on the seq; this commit keeps the
M1 / 850fc37§2 force-writes intact so the protocol stays correct in
the intermediate state.

Mechanical change set:
- `internode.cu:66-79` — `nvl_pack` / `nvl_seq_match` /
  `nvl_unpack_value` now operate on `uint64_t` (seq in high 32 bits,
  value in low 32 bits). Comment block rewritten for the new horizon.
- `internode.cu` (4 sites) — `AsymBuffer<int>` → `AsymBuffer<uint64_t>`
  on the 4 dispatch slot families (allocated separately in
  `dispatch_main_kernel` and `dispatch_grads_main_kernel`) and the
  2 combine slot families (sender side + forwarder/coord side).
- `internode.cu` (~12 sites) — `int raw_*` → `uint64_t raw_*` at every
  slot load. The existing `ld_volatile_global` / `ld_acquire_sys_global`
  overloads for `uint64_t*` carry the type through PTX-level
  `ld.volatile.global.u64` / `ld.acquire.sys.global.u64`. `__shfl_sync`
  dispatches the standard `unsigned long long` overload.
- `utils.cuh:88-100` — add `st_relaxed_sys_global(uint64_t*, uint64_t)`
  and `st_release_sys_global(uint64_t*, uint64_t)` (PTX `.b64`).
- `config.hpp:64` — `(2 * num_rdma_ranks + 3) * sizeof(int)` →
  `* sizeof(uint64_t)` in `get_nvl_buffer_size_hint`. At production
  (32 channels × 8 NVL peers × N=2 RDMA ranks × +3 slack) the NVL
  budget grows by ~7 KB per peer — invisible against the MB-scale
  ring buffers in the same hint.

What stays. M1's iter-entry force-writes at the 5 sites
(dispatch fwd forwarder ~1135, dispatch fwd receiver ~1360,
dispatch_grads forwarder ~2260, dispatch_grads receiver ~2451,
combine sender ~2760, combine coord ~3179) remain. With (a) dead by
construction here and (b) still alive, the force-writes are still
load-bearing for cross-call same-seq aliasing — next commit retires
them in lockstep with the phase-bit change.

Build green via login-node incremental rebuild
(`pixi run --manifest-path evolutionaryscale/pixi.toml bash
build_stream_ep.sh`). GPU validation runs after commit 2 since both
commits land on the same release; intermediate-commit smoke run not
useful (no behavior change versus 850fc37 at low-iter shapes).

Next: phase-bit on `dispatch_seq` (fwd=2N / bwd=2N+1) in
`stream_moe.py` + the bench harness, then delete the 5 force-write
blocks.

+++ csrc/config.hpp
@@ -61,7 +61,12 @@ struct Config {
         const int num_channels = num_sms / 2;
 
         size_t num_bytes = 0;
-        num_bytes += num_channels * num_nvl_ranks * (2 * num_rdma_ranks + 3) * sizeof(int);
+        // NVL gen-stamp slots (`nvl_channel_prefix_start/end`, `nvl_channel_head`,
+        // `nvl_channel_tail` in `internode::dispatch_main_kernel` and the combine
+        // sender/forwarder/coord) are 64-bit `(seq32, value32)` pairs. Sizing
+        // covers the largest user across dispatch and combine, with one slot of
+        // slack.
+        num_bytes += num_channels * num_nvl_ranks * (2 * num_rdma_ranks + 3) * sizeof(uint64_t);
         num_bytes += num_channels * num_nvl_ranks * num_max_nvl_chunked_recv_tokens * hidden_bytes;
 #ifndef DISABLE_NVSHMEM
         num_bytes += num_channels * num_nvl_ranks * num_max_nvl_chunked_recv_tokens * internode::get_source_meta_bytes();
